add test for 'training.py'
