+++ configs/trainer/default.yaml
@@ -1,7 +1,7 @@
 _target_: pytorch_lightning.Trainer
 
 # set `1` to train on GPU, `0` to train on CPU only
-gpus: 1
+gpus: 0
 
 min_epochs: 1
 max_epochs: 10
@@ -9,3 +9,4 @@ max_epochs: 10
 weights_summary: null
 progress_bar_refresh_rate: 5
 resume_from_checkpoint: null
+fast_dev_run: false
