+++ predict_pv_yield/data/dataloader.py
@@ -52,7 +52,11 @@ class NetCDFDataModule(LightningDataModule):
         num_workers: int = 8,
         pin_memory: bool = True,
         data_path="prepared_ML_training_data/v4/",
+        fake_data: bool = False,
     ):
+        """
+        fake_data: random data is created and used instead. This is useful for testing
+        """
         super().__init__()
 
         self.temp_path = temp_path
@@ -62,6 +66,7 @@ class NetCDFDataModule(LightningDataModule):
         self.n_val_data = n_val_data
         self.num_workers = num_workers
         self.pin_memory = pin_memory
+        self.fake_data = fake_data
 
         self.dataloader_config = dict(
             pin_memory=self.pin_memory,
@@ -75,32 +80,74 @@ class NetCDFDataModule(LightningDataModule):
         )
 
     def train_dataloader(self):
-        train_dataset = NetCDFDataset(
-            self.n_train_data,
-            os.path.join(self.data_path, "train"),
-            os.path.join(self.temp_path, "train"),
-            cloud=self.cloud,
-        )
+        if self.fake_data:
+            train_dataset = FakeDataset()
+        else:
+            train_dataset = NetCDFDataset(
+                self.n_train_data,
+                os.path.join(self.data_path, "train"),
+                os.path.join(self.temp_path, "train"),
+                cloud=self.cloud,
+            )
 
         return torch.utils.data.DataLoader(train_dataset, **self.dataloader_config)
 
     def val_dataloader(self):
-        val_dataset = NetCDFDataset(
-            self.n_val_data,
-            os.path.join(self.data_path, "validation"),
-            os.path.join(self.temp_path, "validation"),
-            cloud=self.cloud,
-        )
+        if self.fake_data:
+            val_dataset = FakeDataset()
+        else:
+            val_dataset = NetCDFDataset(
+                self.n_val_data,
+                os.path.join(self.data_path, "validation"),
+                os.path.join(self.temp_path, "validation"),
+                cloud=self.cloud,
+            )
 
         return torch.utils.data.DataLoader(val_dataset, **self.dataloader_config)
 
     def test_dataloader(self):
-        # TODO need to change this to a test folder
-        test_dataset = NetCDFDataset(
-            self.n_val_data,
-            os.path.join(self.data_path, "validation"),
-            os.path.join(self.temp_path, "validation"),
-            cloud=self.cloud,
-        )
+        if self.fake_data:
+            test_dataset = FakeDataset()
+        else:
+            # TODO need to change this to a test folder
+            test_dataset = NetCDFDataset(
+                self.n_val_data,
+                os.path.join(self.data_path, "validation"),
+                os.path.join(self.temp_path, "validation"),
+                cloud=self.cloud,
+            )
 
         return torch.utils.data.DataLoader(test_dataset, **self.dataloader_config)
+
+
+class FakeDataset(torch.utils.data.Dataset):
+    """Fake dataset."""
+
+    def __init__(self, batch_size=32, seq_length=10, width=16, height=16, number_sat_channels=8, length=10):
+        self.batch_size = batch_size
+        self.seq_length = seq_length
+        self.width = width
+        self.height = height
+        self.number_sat_channels = number_sat_channels
+        self.length = length
+
+    def __len__(self):
+        return self.length
+
+    def per_worker_init(self, worker_id: int):
+        pass
+
+    def __getitem__(self, idx):
+
+        x = {
+            "sat_data": torch.randn(
+                self.batch_size, self.seq_length, self.width, self.height, self.number_sat_channels
+            ),
+            "pv_yield": torch.randn(self.batch_size, self.seq_length, 128),
+            "nwp": torch.randn(self.batch_size, 10, self.seq_length, 2, 2),
+        }
+
+        # add a nan
+        x["pv_yield"][0, 0, :] = float("nan")
+
+        return x
