+++ tests/test_utils.py
@@ -1,6 +1,7 @@
 from predict_pv_yield.utils import extras, print_config
 
 from hydra import compose, initialize
+import hydra
 
 import os
 
@@ -11,6 +12,7 @@ def test_utils():
     """
     os.environ["NEPTUNE_API_TOKEN"] = "not_at_token"
 
+    hydra.core.global_hydra.GlobalHydra.instance().clear()
     initialize(config_path="../configs", job_name="test_app")
     config = compose(config_name="config")
